The HeroSearch.tsx component, when called from ProviderList.tsx, received the requested search (the input) sent from LandingPage.tsx.
However, if we deleted the content of the input in HeroSearch.tsx and clicked its "Search" button, nothing happened.

It should have displayed all caregivers, based on the active search icon (elderly, children, cleaning, etc.).

In other words, when the user deletes the content of the input in ProviderList and clicks "Search," it should call onSearch.

Even with an empty string, it should display all caregivers, based on the active search icon.