feat: replace home icon with Cuidamet logo in BottomNav for better branding. In the app's button bar (the one at the bottom), we've placed the Cuidamet icon in a small corner.
We have views where the Cuidamet logo doesn't appear (such as Profile, Chat, Map, Reservations, etc.).

So, we've replaced the Home icon in the app's bottom button bar with a simple Cuidamet icon (without text).

We've kept the size and colors of the previous icon.

This new icon already has a color: green.

In fact, when the Home option is pressed, the icon's color changes to a darker green.

But we've also managed to make it so that when the Home option isn't pressed, it uses the gray color of the other buttons in the button bar. We've also made it so that when the mouse hovers over it, the color changes momentarily as if it were pressed (a hover effect, just like the other icons in the button bar).