The "BottomNav.tsx" component has a button/option called "Inbox".
As soon as the "BottomNav.tsx" component is rendered, a small red circle with a red background is rendered over this "Inbox" option, displaying the number of unread chats.

This red circle will be invisible if the user is not authenticated.

However, if the user is authenticated, it should be rendered or made visible. Also, the circle will become visible the moment the user authenticates, regardless of where they are logged into the app.