The "RoleSelection.tsx" component has two cards/options: - "Looking for a caregiver" - "I want to work"
The issue is that when you click on one of these two options, the "AuthPage.tsx" component renders.
However, it used to render directly in "Login" mode.

Now, when you click on either of these two options, it goes directly into "Register" mode.
And if the "Looking for a caregiver" option is clicked, the "I am a Family Member" option in the "AuthPage.tsx" component is already selected.
And if the "I want to work" option is clicked, the "I am a Caregiver" option in the "AuthPage.tsx" component is already selected.